third_party/blink/web_tests/fast/spatial-navigation/snav-hidden-iframe-zero-size-expected.txt

This is link_1.


This is 
an iframe with size 0.


This is link_2.

PASS successfullyParsed is true

TEST COMPLETE

PASS gFocusedDocument.activeElement.getAttribute("id") is "end"

This test is to test that an iframe with size 0 does not get focus, and neither do its descendants.
